Data.Double.Conversion.Convertable

Description

Documentation

class (RealFloat a, IsString b) => Convertable a b where

Type class for floating data types, that cen be converted, using double-conversion library

Default instanced convert input to Double and then make Bytestring Builder from it.

list of functions :

toExponential: Compute a representation in exponential format with the requested number of digits after the decimal point. The last emitted digit is rounded. If -1 digits are requested, then the shortest exponential representation is computed.

toPrecision: Compute precision leading digits of the given value either in exponential or decimal format. The last computed digit is rounded.

toFixed: Compute a decimal representation with a fixed number of digits after the decimal point. The last emitted digit is rounded.

toShortest: Compute the shortest string of digits that correctly represent the input number.

Conversion to text is twice faster than conversion to bytestring Conversion to text via Builder (both in the in case of bytestring and text) in case of single number is much slower, than to text or bytestring directly. (2-3x) But conversion large amount of numbers to text via Builder is much faster than directly (up to 50x). Conversion to text via text builder is a little slower, then via bytestring builder

Minimal complete definition

Nothing

Methods

toExponential :: Int -> a -> b

default toExponential :: b ~ Builder => Int -> a -> b

toPrecision :: Int -> a -> b

default toPrecision :: b ~ Builder => Int -> a -> b

toFixed :: Int -> a -> b

default toFixed :: b ~ Builder => Int -> a -> b

toShortest :: a -> b

default toShortest :: b ~ Builder => a -> b