Data.Double.Conversion.Text

Description

This module left now only for compatibility and should not be used in new projects. Please, use Convertable type class from Data.Double.Conversion.Convertable

Fast, efficient support for converting between double precision floating point values and text.

These functions are about 30 times faster than the default show implementation for the Double type.

Documentation

toExponential :: Int -> Double -> Text

Compute a representation in exponential format with the requested number of digits after the decimal point. The last emitted digit is rounded. If -1 digits are requested, then the shortest exponential representation is computed.

toFixed :: Int -> Double -> Text

Compute a decimal representation with a fixed number of digits after the decimal point. The last emitted digit is rounded.

toPrecision :: Int -> Double -> Text

Compute precision leading digits of the given value either in exponential or decimal format. The last computed digit is rounded.

toShortest :: Double -> Text

Compute the shortest string of digits that correctly represent the input number.